Improve cake-autorate_setup script

+++ cake-autorate_setup.sh
@@ -7,10 +7,14 @@
 # Set correctness options
 set -eu
 
+# Setup dependencies to check for
+DEPENDENCIES="jsonfilter uclient-fetch tar grep"
+
 # Set up remote locations and branch
-SRC_DIR="https://github.com/lynxthecat/cake-autorate/archive/refs/heads/"
-DOC_URL="https://github.com/lynxthecat/CAKE-autorate#installation-on-openwrt"
 BRANCH="master"
+SRC_DIR="https://github.com/lynxthecat/cake-autorate/archive/"
+API_URL="https://api.github.com/repos/lynxthecat/cake-autorate/commits/${BRANCH}"
+DOC_URL="https://github.com/lynxthecat/CAKE-autorate#installation-on-openwrt"
 
 # Check if an instance of cake-autorate is already running and exit if so
 if [ -d /var/run/cake-autorate ]; then
@@ -29,6 +33,15 @@ if [ "$(opkg list-installed | grep -Ec '^(bash|iputils-ping|fping) ')" -ne 3 ]; 
 	opkg install bash iputils-ping fping
 fi
 
+exit_now=0
+for dep in ${DEPENDENCIES}; do
+	if ! type "${dep}" >/dev/null 2>&1; then
+		printf >&2 "%s is required, please install it and rerun the script!\n" "${dep}"
+		exit_now=1
+	fi
+done
+[ ${exit_now} -ge 1 ] && exit ${exit_now}
+
 # Set up CAKE-autorate files
 # cd to the /root directory
 cd /root/ || exit
@@ -39,13 +52,21 @@ cd /root/ || exit
 # cd into it
 cd cake-autorate/ || exit
 
+# Get the latest commit to download
+commit=$(uclient-fetch -qO- "${API_URL}" | jsonfilter -e @.sha)
+if [ -z "${commit:-}" ];
+then
+	printf >&2 "Invalid operation occurred, commit variable should not be empty"
+	exit 1
+fi
+
 printf "Installing cake-autorate in /root/cake-autorate...\n"
 
 # Download the files to a temporary directory, so we can move them to the cake-autorate directory
 tmp=$(mktemp -d)
 trap 'rm -rf "${tmp}"' EXIT INT TERM
-wget -qO- "${SRC_DIR}/${BRANCH}.tar.gz" | tar -xozf - -C "${tmp}"
-mv "${tmp}/cake-autorate-${BRANCH}"/* "${tmp}"
+uclient-fetch -qO- "${SRC_DIR}/${commit}.tar.gz" | tar -xozf - -C "${tmp}"
+mv "${tmp}/cake-autorate-"*/* "${tmp}"
 
 # Check if a configuration file exists, and ask whether to keep it
 editmsg="\nNow edit the cake-autorate_config.primary.sh file as described in:\n   ${DOC_URL}"
@@ -58,7 +79,7 @@ if [ -f cake-autorate_config.primary.sh ]; then
 		editmsg="Using prior configuration"
 		mv "${tmp}/cake-autorate_config.primary.sh" cake-autorate_config.primary.sh.new
 	fi
-else 
+else
 	mv "${tmp}/cake-autorate_config.primary.sh" cake-autorate_config.primary.sh
 fi
 
@@ -69,6 +90,13 @@ for file in ${files}; do
 	mv "${tmp}/${file}" "${file}"
 done
 
+# Get version and generate a file containing version information
+version=$(grep -m 1 ^cake_autorate_version= /root/cake-autorate/cake-autorate.sh | cut -d= -f2 | cut -d'"' -f2)
+cat > version.txt <<-EOF
+	version = ${version}
+	commit = ${commit}
+EOF
+
 # Also copy over the service file but DO NOT ACTIVATE IT
 mv "${tmp}/cake-autorate" /etc/init.d/
 chmod +x /etc/init.d/cake-autorate
@@ -77,8 +105,7 @@ chmod +x /etc/init.d/cake-autorate
 # shellcheck disable=SC2059
 printf "${editmsg}\n"
 
-# shellcheck disable=SC2312
-printf '\n%s\n\n' "$(grep -m 1 ^cake_autorate_version= /root/cake-autorate/cake-autorate.sh | cut -d= -f2 | cut -d'"' -f2) successfully installed, but not yet running"
+printf '\n%s\n\n' "${version} successfully installed, but not yet running"
 printf '%s\n' "Start the software manually with:"
 printf '%s\n' "   cd /root/cake-autorate; ./cake-autorate.sh"
 printf '%s\n' "Run as a service with:"
